turtle: clear turtle_ready flag when quad is right side up
Currently the turtle_ready flag is only cleared if turtle action succeeds
if the action fails and you manually flip the drone over the turtle_ready
flag is stuck on. If the initial motor direction fails, the next attempt
to set the flag when armed will not run with flag set. This clears the flag
to allow the motor direction change to work.